Fix hordes are blocked by non func_brush brushes
Hordes and wanderers appeared to be blocked by clip or npc_clip brushes which
do not actually block the alien movement. Mostly noticeable on Rydberg Reactor
finale, where if players are blocking the finale area, hordes and wanderers
don't come from the south, but they should. This now should be fixed.