Fix aliens are stuck indefinitely around dead marine
This fix makes sure that dead marine is not considered as the one
who sees aliens, thus preventing them from going into sleep state.
This fixes the bug: When marine dies to a horde and other marines are far
enough for horde to see them, the horde gets stuck in non-sleeping state
and thus not decreasing the ammount of non-sleeping aliens. This leads to
no hordes and wanderers being spawned at all and can be exploited on hard
challenges and maps like Survival Desert.