Fix 64 fps server limit on newer windows versions
Description
This fixes an issue for the dedicated server on Windows 10 build 2004 or newer.

Background
Microsoft has deprecated timers in this and later builds. In the past, it was possible to claim a high resolution timer (1ms) for games. Later, an external utility named 'Source fps booster' was needed to achieve the same. Since Windows build 2004, only the program itself can claim this. Without the fix, the dedicated server would be stuck around 64 fps. With this fix, it goes up to 500 (or 1000) as it was before, or fps_max if set.

Working
This patch claims a high resolution timer on behalf of the main thread as soon as the first player connects (as in: when the server awakes from hibernation).

This way, idle servers won't consume any more energy or performance than strictly needed.