Make 'Keep Mouse Inside Window' setting visible for fullscreen mode too
Keep mouse inside window seems to not only work for windowed, but also in fullscreen (when multiple monitors are enabled).
With this setting disabled in fullscreen, the mouse wanders to the secondary screen during gameplay (and alt tabs upon mouse click).
How to reproduce:
 1 disable the keep inside window option in windowed mode;
 2 switch to fullscreen
 3 play a game and notice the cursor can go outside the primary monitor (on multiple monitors).
Solution is simple: make the option in fullscreen available too.