Prevent AI medic from running too far to heal a marine
The problem with AI medics is that they try to run half through the map
to heal a marine and die along the way. This happens when team splits, on maps
such as SynTek Residential.

Added rd_marine_heal_range_max ConVar to control the distance(a radius actually)
for AI medic to search for heal targets. 512 units is the default range. The
ConVar however accepts the square of this range, so it should be set to
512*512=262144.