Remove grenade interaction with water.
Previously, grenades more than 18 inches (24 hammer units) under the surface of water would not affect anything not touching water and grenades that were not that deeply submerged into water would not affect anything with its eyes under water.

There's not much deep water in the game, including in custom campaigns, and most of the time, deep water just kills whatever is in it automatically.

Additionally, many creatures in Alien Swarm have eyes that are not 18 inches above their feet, so the water interaction with explosions is at best confusing.