Change last fight in res3miningcamp
Changed how aliens attack the generator. The issue with a stuck unkillable alien
that damages the generator and fails the mission should be fixed.
However players can now also damage the generator. So care should be taken when firing explosive weapons near it.
Restored original lighting, water and projector. Map is now darker, as the
original author planned. But also looks more nice.