allow multiple status effects to be layered on a single alien
this isn't perfect and animation is going to be tied to the highest-motion status effect applied to the alien. also, this is exponentially expensive. with three status effects, this raises our texture count from 3 to 7, but if we ever add another status effect to this system it'll raise the texture count to 15.

should have minimal runtime impact on performance as the new textures are all relatively small and the complexity of the shader is not changing.