New RGB Blending for Chat Text
Originally, I was just messing around with this code not planning on uploading it, but peeps seemed to like it so I Improved it and added more functionality.

This new addition adds blending for chat text. It works similar to the previous RGB system and uses two of the same 0x08 characters to send data to the client. This is followed by an encoded blend type to tell the client which blend mode we are using. These new blending functions would allow a new colorful blend of colors for modders and server owners to use.

The first is the basic blend that blends two sets of RGB colors. The blend starts at the point after the encoded sequence finishes and ends after (either A. The end of the text B. Another color char is detected.)

Example:
ClientPrint(null, 3, TextColorBlend(255, 155, 0, 255, 255, 0) + "This is a test to show blended text using new color blend function. As you can see the text blends from color1 to color2 as it gets closer to the tail end of the text.");

The next is an inversion blend that blends a single RGB input to it's inverted color. Works the same way as normal blend only it doesn't need to send extra 3 bytes for second RGB input.

Example:

ClientPrint(null, 3, TextColorBlendInvert(255, 255, 125) + "This is a test to show inverted blend text. It blends into the inverted color for the supplied RGB input.");

The next is a blend of 3 colors. Works the same only has an additional RGB color.

Example:

ClientPrint(null, 3, TextColorBlend3(255, 55, 55, 255, 255, 255, 55, 55, 255) + "This is a test to show the 3 color blend text. It blends into the 3 colors for the supplied RGB inputs.");

The next is a color blend cycle. The color blend cycle's first argument is the length of a transition followed by two RGB inputs. The blend cycle allows repeating a color pattern. The pattern will jump to color 1 after it reaches the end of it's provided length.

Example:
ClientPrint(null, 3, TextColorBlendCycle(5, 122, 122, 255, 155, 155, 155) + "This is a test to show the color cycle blend text. It cycles the blend through color 1 to color 2 using the inputted length to control how long the blend is.");

The last is a color blend smooth cycle. It works the same as normal blend cycle only it smoothly returns to color 1.

Example:
ClientPrint(null, 3, TextColorBlendSmoothCycle(25, 255, 155, 155, 155, 155, 255) + "This is a test to show the color smooth cycle blend text. It cycles a smooth blend through color 1 to color 2 and back to color 1 using the inputted length to control how long the blend is.");