Deima Bug Fix
I got stuck on bridge with bloodhound and we couldn't progress so I looked into it.

I believe it was caused by a marine skipping past the trigger.

I extended the trigger across the whole edge of the bridge so it should trigger each time. The reset trigger delay was dropped to 0.1 from 1.0 just in case that was also a factor.