Partial Steam Input implementation
This works well so far, but we have no way of doing analog inputs yet.
As-is, enabling the release flag and adding an action manifest will
cause the Steam Input Configurator to no longer present "As Joystick" as
an option for stick input, meaning players using controller would no
longer be able to move or aim. Not optimal. The glyph code works as-is,
so it is enabled without a release flag.