Remove difficulty-specific leaderboards as a concept.
We currently have 9496 leaderboards. There is a hard limit of 10000
leaderboards per game in the Steam API.

Of the leaderboards that currently exist:

- There are 665 leaderboards for no challenge.
- There are 1283 difficulty-specific leaderboards for no challenge.
- There are 1410 leaderboards for official challenges.
- There are 2502 difficulty-specific leaderboards for official challenges.
- There are 1972 leaderboards for workshop challenges.
- There are 1646 difficulty-specific leaderboards for workshop challenges.
- The remaining 2 leaderboards are for Heroes of the Interstellar Armed Forces,
  but we could probably get that down to 1.

This commit removes the 1283+2502+1646 leaderboards (57%) that contain
mostly duplicated information and will allow us to free up space for
future maps added to the game by the community. Otherwise, I expect us
to completely run out of leaderboard slots before the end of the first
annual mapping competition.

We can look into removing some of the whitelist entries for workshop
challenges that are less popular than expected or that have been
superseded by other challenges in the future in order to free up more
space.

The difficulty-specific leaderboards will remain in existence until the
anniversary update as the non-beta version of the game does not know
about this change. That also gives us time to change our mind on this
decision if we come up with a different solution before then.