VScript: player.ShowMenu(options, timeout, text)
Response comes back as a menuselect command; catch with the
UserConsoleCommand hook.

options is a bitfield of valid options, the sum of powers of two for
keys that will be accepted. F1-F10 are 1, 2, 4, 8, 16, 32, 64, 128, 256, 512.
1023 means all ten keys will be accepted, 7 means F1-F3 only, 519 means
F1, F2, F3, and F10. Pass 0 for options to hide the current menu.

duration is an integer from -1 to 127, the number of seconds to display
this menu for. -1 means no timeout.

text is a specially formatted string of no more than 511 characters.
Example: "->1. First option\n->2. Second option\n\n->0. Last option\n"