new map voting system
now all voters count towards the vote fraction instead of all players on the server.

in the previous system not voting is treated as voting no.

new system will not factor in afk people and people who don't care about next map and want to leave it for others to vote