Add convars to lock the difficulty settings for a lobby. Suggested by 法克*尤尔*马泽尔.
These cannot be easily set by a challenge (point_servercommand can still
do it, but the easier methods of changing a convar will not work for
these new cvars or for hostname or sv_tags)

Challenges that set onslaught or hardcore friendly fire still prevent
changing those settings from the lobby as before. This lock only affects
the lobby UI and the commands the lobby UI calls; other methods of
changing the difficulty settings still work.

Open to suggestions for additional convars that should be marked as
off-limits to challenges, as well as anything else, as always.