increase default rate and add fcvar_archive to rate.
This will not affect bandwidth requirements when the game isn't lagging,
but when the game's connection is maxed out, this increases the maximum
upload bandwidth for a server with 32 players using only default
settings from 2.56 Mbps to 25.6 Mbps. This is still affected by users
with modified settings, servers with modified settings, etc., and this
bandwidth increase is only the maximum; again, the amount of bandwidth
needed by the game is not changing here, only the maximum amount the
game is willing to use with default settings.