Temp banned dm_residential
Until dm_residential is compatible with Infection, this map is locked from being playable in Infection.

Editing a map for this mode takes a very long time. dm_deima, for example, took a very long time to get right. I would need to consider patching all the game-breaking exploits and try to balance out all the areas that are not easily accessible so they do not give newer players major disadvantages.